SDIT-743 Refactor integration tests
